+++ src/unstract/sdk/adapters/x2text/no_op/src/__init__.py
@@ -1,9 +1,9 @@
-from .no_op_x2text_adapter import NoOpX2TextAdapter
+from unstract.sdk.adapters.x2text.no_op.src.no_op_x2text import NoOpX2Text
 
 metadata = {
-    "name": NoOpX2TextAdapter.__name__,
+    "name": NoOpX2Text.__name__,
     "version": "1.0.0",
-    "adapter": NoOpX2TextAdapter,
-    "description": "NoOpX2TextAdapter",
+    "adapter": NoOpX2Text,
+    "description": "NoOpX2Text",
     "is_active": True,
 }
